+++ Show-SolutionDgml.ps1
@@ -1,6 +1,6 @@
 
 <#PSScriptInfo
-.VERSION 1.0.2
+.VERSION 1.1.0
 .AUTHOR Roman Kuzmin
 .COPYRIGHT (c) Roman Kuzmin
 .TAGS DGML, solution, project, graph
@@ -26,6 +26,10 @@
 		Specifies the solution path. If it is omitted or empty then the *.sln
 		file in the current location is used, there must be one such file.
 
+.Parameter Exclude
+		Specifies the projects to exclude.
+		Use project names without extensions.
+
 .Parameter JustProject
 		Tells to show just references defined in project files and ignore build
 		order dependencies in the solution.
@@ -41,6 +45,7 @@
 [CmdletBinding()]
 param(
 	[string]$SolutionPath,
+	[string[]]$Exclude,
 	[switch]$JustProject,
 	[switch]$JustSolution
 )
@@ -163,7 +168,15 @@ $links = $doc.AppendChild($xml.CreateElement('Links'))
 
 $ns = @{x = 'http://schemas.microsoft.com/developer/msbuild/2003'}
 
+function Test-Exclude($Name) {
+	foreach($_ in $Exclude) {
+		if ($_ -eq $Name) {return $true}
+	}
+}
+
 foreach($project in $map.Values) {
+	if (Test-Exclude ($project.name)) {continue}
+
 	$node = $nodes.AppendChild($xml.CreateElement('Node'))
 	$node.SetAttribute('Id', $project.name)
 	$node.SetAttribute('Path', $project.path)
@@ -174,6 +187,8 @@ foreach($project in $map.Values) {
 		foreach($match in $reMatchProjectDependency.Matches($Matches[1])) {
 			$id = $match.Groups[1].Value
 			$project2 = $map[$id]
+			if (Test-Exclude ($project2.name)) {continue}
+
 			$link = $links.AppendChild($xml.CreateElement('Link'))
 			$link.SetAttribute('Source', $project.name)
 			$link.SetAttribute('Target', $project2.name)
@@ -203,6 +218,7 @@ foreach($project in $map.Values) {
 		# write project links
 		foreach ($reference in $references) {
 			$name2 = [System.IO.Path]::GetFileName($reference.Node.'#text')
+
 			$project2 = @(
 				foreach($_ in $map.Values) {
 					if ([System.IO.Path]::GetFileName($_.path) -eq $name2) {
@@ -210,8 +226,20 @@ foreach($project in $map.Values) {
 					}
 				}
 			)
-			if ($project2.Count -eq 0) {throw "Cannot find '$name2' in the solution."}
+
+			# When a project is removed from a solution its references are not
+			# removed from unloaded projects, so missing links are possible.
+			if ($project2.Count -eq 0) {
+				Write-Warning "Cannot find '$name2' referenced by '$projectPath' in the solution."
+				continue
+			}
+
+			# In theory, we may have several projects with the same name but
+			# with different extension of path. This is weird, fail for now.
 			if ($project2.Count -ge 2) {throw "Too many '$name2' in the solution."}
+
+			if (Test-Exclude ($project2[0].name)) {continue}
+
 			$link = $links.AppendChild($xml.CreateElement('Link'))
 			$link.SetAttribute('Source', $project.name)
 			$link.SetAttribute('Target', $project2[0].name)